requirements.txt
streamlit
langchain
langchain_core
langchain_community
python-dotenv
langchain-groq
plotly
pandas
pymysql
langchain_openai
mysql-connector-python
faker
google-search-results
faiss